本次更新了商家热力榜查询功能，修改了部分mapper.xml字段类型和相应的entity实体类属性类型，去除了Result的冗余属性和构造方法，暂时关闭登录拦截器以方便后续功能的添加。

+++ src/main/java/com/taotao/entity/Merchant.java
@@ -37,7 +37,7 @@ public class Merchant implements Serializable {
      * 粉丝数量
      */
     @TableField(value = "fansCount")
-    private Integer fansCount;
+    private Long fansCount;
 
     /**
      * 逻辑删除
